# openapi.yaml
openapi: 3.0.3
info:
  title: GIA — Gemini Intelligent Assistant API
  description: |
    Production-grade API definition for GIA.
    Orchestrates memory stores, document uploads, semantic searches, RAG pipelines, and multi-LLM workflows with Redis session storage and token revocation capabilities.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local development backend server
paths:
  /auth/signup:
    post:
      summary: User Registration
      description: Create a new user account and set up a stateful session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: developer@gia.ai
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: secure_pass_123
                name:
                  type: string
                  minLength: 1
                  example: Alice Developer
      responses:
        '211':
          description: Created successfully (maps to 201)
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session_id=abc123xyz; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: Opaque session ID (returned for API client compatibility)
                    example: 7c57f7dfd311910cf9cf6a3bcf4489a7f34c2b9a764d23bb5e468da40177fc22
        '400':
          description: Validation error or duplicate email
  /auth/login:
    post:
      summary: User Authentication
      description: Validate credentials and initialize a Redis session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: developer@gia.ai
                password:
                  type: string
                  format: password
                  example: secure_pass_123
      responses:
        '200':
          description: Logged in successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session_id=abc123xyz; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: Opaque session ID
                    example: 7c57f7dfd311910cf9cf6a3bcf4489a7f34c2b9a764d23bb5e468da40177fc22
        '400':
          description: Validation failed
        '401':
          description: Invalid email or password
  /auth/me:
    get:
      summary: Get User Profile
      description: Retrieve user details from the active session.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Session valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Missing or expired session
  /auth/logout:
    post:
      summary: Session Sign Out
      description: Revoke the active session key from Redis and clear cookies.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Successfully signed out
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session_id=; Path=/; Max-Age=0; Expires=Thu, 01 Jan 1970 00:00:00 GMT
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
  /auth/revoke-all:
    post:
      summary: Revoke All Sessions
      description: Invalidate all active session IDs for the authenticated user (e.g., on key compromise).
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: All sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
...
  /memories:
    get:
      summary: List/Keyword Search Memories
      description: Retrieve user memories filtered by keyword filters.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Optional case-insensitive keyword search
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Memories retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
    post:
      summary: Create Memory
      description: Manually insert a preference or fact memory record.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - type
                - confidence
              properties:
                content:
                  type: string
                  minLength: 1
                  example: User prefers Python 3.12 over Node.js
                type:
                  type: string
                  enum: [user_preference, explicit_note, conversation_context]
                  example: user_preference
                importance:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 5
                  example: 8
                confidence:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  example: 0.95
                metadata:
                  type: object
                  example: { "source": "settings_panel" }
      responses:
        '201':
          description: Memory saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  memory:
                    $ref: '#/components/schemas/Memory'
  /memories/search:
    get:
      summary: Semantic Search Memories
      description: Search memories using vector cosine similarity.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Search text query
        - name: semantic
          in: query
          schema:
            type: boolean
            default: true
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Match records found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
  /memories/{id}:
    put:
      summary: Update Memory
      description: Update memory content or rating parameters.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  minLength: 1
                  example: User prefers Python 3.12 and Go.
                importance:
                  type: integer
                  minimum: 1
                  maximum: 10
                type:
                  type: string
                  enum: [user_preference, explicit_note, conversation_context]
                confidence:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                metadata:
                  type: object
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  memory:
                    $ref: '#/components/schemas/Memory'
        '403':
          description: Access denied
    delete:
      summary: Delete Memory
      description: Remove memory record.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
  /memories/extract:
    post:
      summary: Extract Memories from Conversation
      description: Ask LLM to parse a conversation thread, extract facts, and save them.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - conversationId
                - content
              properties:
                conversationId:
                  type: string
                  format: uuid
                  example: e6d092d6-47b2-4d43-aa9d-29bcbb3f374c
                content:
                  type: string
                  description: Dialog context string to extract facts from
                  example: "Alice mentioned today that she is learning Rust."
      responses:
        '200':
          description: Extraction complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d2d092d6-47b2-4d43-aa9d-29bcbb3f374c
        email:
          type: string
          format: email
          example: developer@gia.ai
        name:
          type: string
          example: Alice Developer
...
    Memory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: f9d092d6-47b2-4d43-aa9d-29bcbb3f374c
        user_id:
          type: string
          format: uuid
          example: d2d092d6-47b2-4d43-aa9d-29bcbb3f374c
        type:
          type: string
          enum: [user_preference, explicit_note, conversation_context]
          example: user_preference
        content:
          type: string
          example: User prefers Python 3.12 over Node.js
        importance:
          type: integer
          example: 8
        confidence:
          type: number
          example: 0.95
        metadata:
          type: object
          nullable: true
          example: { "source": "settings_panel" }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...